chore(company): handle duplicate emails
